Introduce fixated pages extension
