+++ src/pages/api/chicken.api.ts
@@ -1,0 +1,13 @@
+export default function handler(req: any, res: any) {
+  res.status(200).json({
+    pacow: `
+    \n     __//
+    \ncf  /.__.\
+    \n    \ \/ /
+    \n '__/    \
+    \n \-      )
+    \n   \_____/
+    \n_____|_|____
+    \n     " "     `,
+  })
+}
